refactor: rename component name and place on components folder
